+++ Assets/Font/demo-files/demo.css
@@ -149,4 +149,10 @@ p {
 .fs1 {
   font-size: 32px;
 }
+.fs2 {
+  font-size: 24px;
+}
+.fs3 {
+  font-size: 32px;
+}
