+++ Assets/Font/style.css
@@ -1,10 +1,10 @@
 @font-face {
   font-family: 'IcoMoon';
-  src:  url('fonts/IcoMoon.eot?7rjkf5');
-  src:  url('fonts/IcoMoon.eot?7rjkf5#iefix') format('embedded-opentype'),
-    url('fonts/IcoMoon.ttf?7rjkf5') format('truetype'),
-    url('fonts/IcoMoon.woff?7rjkf5') format('woff'),
-    url('fonts/IcoMoon.svg?7rjkf5#IcoMoon') format('svg');
+  src:  url('fonts/IcoMoon.eot?1aydyt');
+  src:  url('fonts/IcoMoon.eot?1aydyt#iefix') format('embedded-opentype'),
+    url('fonts/IcoMoon.ttf?1aydyt') format('truetype'),
+    url('fonts/IcoMoon.woff?1aydyt') format('woff'),
+    url('fonts/IcoMoon.svg?1aydyt#IcoMoon') format('svg');
   font-weight: normal;
   font-style: normal;
   font-display: block;
@@ -25,6 +25,18 @@
   -moz-osx-font-smoothing: grayscale;
 }
 
+.icon-equalizer:before {
+  content: "\e90e";
+}
+.icon-share:before {
+  content: "\e90f";
+}
+.icon-content_copy:before {
+  content: "\e910";
+}
+.icon-more_vert:before {
+  content: "\e911";
+}
 .icon-wand:before {
   content: "\e900";
 }
