+++ Playground/Playground/Features/CssPreviewer/CssPreviewerViewModel.cs
@@ -2,7 +2,6 @@
 using MagicGradients.Parser;
 using MagicGradients.Xaml;
 using Playground.Data.Repositories;
-using Playground.ViewModels;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
@@ -11,126 +10,11 @@ using Xamarin.Forms;
 
 namespace Playground.Features.CssPreviewer
 {
-    [QueryProperty("Id", "id")]
-    [QueryProperty("Data", "data")]
-    public class CssPreviewerBase : ObservableObject
-    {
-        private readonly IGradientRepository _gradientRepository;
-
-        private string _id;
-        public string Id
-        {
-            get => _id;
-            set
-            {
-                _id = value;
-                LoadCssCodeFromGallery(int.Parse(_id));
-            }
-        }
-
-        private string _data;
-        public string Data
-        {
-            get => _data;
-            set
-            {
-                _data = value;
-                ParseData(_data);
-            }
-        }
-
-        private string _cssCode;
-        public string CssCode
-        {
-            get => _cssCode;
-            set => SetProperty(ref _cssCode, value, () =>
-            {
-                if (IsHotReload)
-                    UpdateGradientSource();
-            });
-        }
-
-        private string _cssSize;
-        public string CssSize
-        {
-            get => _cssSize;
-            set => SetProperty(ref _cssSize, value, () =>
-            {
-                if (IsHotReload)
-                    UpdateGradientSize();
-            });
-        }
-
-        private string _cssRepeat;
-        public string CssRepeat
-        {
-            get => _cssRepeat;
-            set => SetProperty(ref _cssRepeat, value, () =>
-            {
-                if (IsHotReload)
-                    UpdateGradientRepeat();
-            });
-        }
-
-        private bool _isHotReload = true;
-        public bool IsHotReload
-        {
-            get => _isHotReload;
-            set => SetProperty(ref _isHotReload, value);
-        }
-
-        public CssPreviewerBase(IGradientRepository gradientRepository)
-        {
-            _gradientRepository = gradientRepository;
-        }
-
-        private void LoadCssCodeFromGallery(int id)
-        {
-            var gradient = _gradientRepository.GetById(id);
-
-            if (gradient == null)
-                return;
-
-            CssCode = gradient.Stylesheet;
-            CssSize = gradient.Size;
-        }
-
-        private void ParseData(string data)
-        {
-            var parts = Uri.UnescapeDataString(data).Split(new [] { ';' }, StringSplitOptions.RemoveEmptyEntries);
-
-            if (parts.Length >= 1)
-                CssCode = parts[0];
-
-            if (parts.Length >= 2)
-                CssSize = parts[1];
-
-            if (parts.Length >= 3)
-                CssRepeat = parts[2];
-        }
-
-        protected virtual void UpdateGradientSource()
-        {
-
-        }
-
-        protected virtual void UpdateGradientSize()
-        {
-
-        }
-
-        protected virtual void UpdateGradientRepeat()
-        {
-
-        }
-    }
-    
     public class CssPreviewerViewModel : CssPreviewerBase
     {
         private readonly DimensionsTypeConverter _dimensionsConverter;
         private readonly BackgroundRepeatTypeConverter _repeatConverter;
         private CssSnippet[] _snippets;
-
         
         public GradientCollection GradientSource { get; set; }
 
@@ -158,6 +42,13 @@ namespace Playground.Features.CssPreviewer
                 () => RaisePropertyChanged(nameof(IsMessageVisible)));
         }
 
+        private bool _isHotReload = true;
+        public bool IsHotReload
+        {
+            get => _isHotReload;
+            set => SetProperty(ref _isHotReload, value);
+        }
+
         public ICommand ClearCommand { get; }
         public ICommand ShowSnippetsCommand { get; }
         public ICommand RefreshCommand { get; }
@@ -183,7 +74,22 @@ namespace Playground.Features.CssPreviewer
             UpdateGradientSource();
         }
 
-        protected override void UpdateGradientSource()
+        protected override void OnPropertyChanged(string propertyName)
+        {
+            if (!IsHotReload)
+                return;
+
+            if(propertyName == nameof(CssCode))
+                UpdateGradientSource();
+
+            if (propertyName == nameof(CssSize))
+                UpdateGradientSize();
+
+            if (propertyName == nameof(CssRepeat))
+                UpdateGradientRepeat();
+        }
+
+        private void UpdateGradientSource()
         {
             Message = string.Empty;
 
@@ -201,7 +107,7 @@ namespace Playground.Features.CssPreviewer
             }
         }
 
-        protected override void UpdateGradientSize()
+        private void UpdateGradientSize()
         {
             Message = string.Empty;
 
@@ -215,7 +121,7 @@ namespace Playground.Features.CssPreviewer
             }
         }
 
-        protected override void UpdateGradientRepeat()
+        private void UpdateGradientRepeat()
         {
             Message = string.Empty;
